LPS-54223 move DataSample impls to module
