+++ modules/portal/portal-monitoring/src/com/liferay/portal/monitoring/internal/statistics/service/ServiceRequestDataSample.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.monitoring.statistics.service;
+package com.liferay.portal.monitoring.internal.statistics.service;
 
 import com.liferay.portal.kernel.monitoring.MethodSignature;
 import com.liferay.portal.monitoring.MonitorNames;
-import com.liferay.portal.monitoring.statistics.BaseDataSample;
+import com.liferay.portal.monitoring.internal.statistics.BaseDataSample;
 
 import java.lang.reflect.Method;
